requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
tensorflow==2.15.0
ta==0.10.1 
numpy
matplotlib
scikit-learn
keras       